RC-1400 move the pkg/config/remote package and /pkg/remoteconfig/state module to a new RC module named /pkg/remoteconfig
