+++ cmd/agent/api/grpc.go
@@ -13,6 +13,7 @@ package api
 import (
 	"context"
 	"fmt"
+	remoteconfig "github.com/DataDog/datadog-agent/pkg/remoteconfig/service"
 	"time"
 
 	workloadmetaServer "github.com/DataDog/datadog-agent/comp/core/workloadmeta/server"
@@ -23,7 +24,6 @@ import (
 
 	dsdReplay "github.com/DataDog/datadog-agent/comp/dogstatsd/replay"
 	dogstatsdServer "github.com/DataDog/datadog-agent/comp/dogstatsd/server"
-	remoteconfig "github.com/DataDog/datadog-agent/pkg/config/remote/service"
 	pb "github.com/DataDog/datadog-agent/pkg/proto/pbgo/core"
 	"github.com/DataDog/datadog-agent/pkg/tagger"
 	"github.com/DataDog/datadog-agent/pkg/tagger/replay"
