+++ pkg/remoteconfig/go.mod
@@ -1,0 +1,88 @@
+module github.com/DataDog/datadog-agent/pkg/remoteconfig
+
+go 1.20
+
+require (
+	github.com/DataDog/datadog-agent/pkg/proto v0.50.0-rc.4
+	github.com/DataDog/datadog-agent/pkg/telemetry v0.50.0-rc.4
+	github.com/DataDog/datadog-agent/pkg/util/backoff v0.50.0-rc.4
+	github.com/DataDog/datadog-agent/pkg/util/http v0.50.0-rc.4
+	github.com/DataDog/datadog-agent/pkg/util/log v0.50.0-rc.4
+	github.com/DataDog/datadog-agent/pkg/version v0.50.0-rc.4
+	github.com/DataDog/go-tuf v1.0.2-0.5.2
+	github.com/Masterminds/semver v1.5.0
+	github.com/benbjohnson/clock v1.3.0
+	github.com/pkg/errors v0.9.1
+	github.com/secure-systems-lab/go-securesystemslib v0.7.0
+	github.com/stretchr/testify v1.8.4
+	go.etcd.io/bbolt v1.3.7
+	google.golang.org/grpc v1.59.0
+	google.golang.org/protobuf v1.31.0
+)
+
+// These replaces are for modules local to the datadog-agent repo.
+replace (
+	github.com/DataDog/datadog-agent => ../..
+	github.com/DataDog/datadog-agent/pkg/proto => ../proto
+	github.com/DataDog/datadog-agent/pkg/telemetry => ../telemetry/
+	github.com/DataDog/datadog-agent/pkg/util/backoff => ../util/backoff
+	github.com/DataDog/datadog-agent/pkg/util/http => ../util/http/
+	github.com/DataDog/datadog-agent/pkg/util/log => ../util/log
+	github.com/DataDog/datadog-agent/pkg/version => ../version
+)
+
+require (
+	github.com/DataDog/datadog-agent/comp/core/telemetry v0.50.0-rc.4 // indirect
+	github.com/DataDog/datadog-agent/pkg/config/model v0.50.0-rc.4 // indirect
+	github.com/DataDog/datadog-agent/pkg/util/fxutil v0.50.0-rc.4 // indirect
+	github.com/DataDog/datadog-agent/pkg/util/scrubber v0.50.0-rc.4 // indirect
+	github.com/DataDog/viper v1.12.0 // indirect
+	github.com/beorn7/perks v1.0.1 // indirect
+	github.com/cespare/xxhash/v2 v2.2.0 // indirect
+	github.com/cihub/seelog v0.0.0-20170130134532-f561c5e57575 // indirect
+	github.com/davecgh/go-spew v1.1.2-0.20180830191138-d8f796af33cc // indirect
+	github.com/fsnotify/fsnotify v1.7.0 // indirect
+	github.com/go-logr/logr v1.3.0 // indirect
+	github.com/go-logr/stdr v1.2.2 // indirect
+	github.com/golang/protobuf v1.5.3 // indirect
+	github.com/google/go-cmp v0.6.0 // indirect
+	github.com/google/gofuzz v1.2.0 // indirect
+	github.com/grpc-ecosystem/grpc-gateway v1.16.0 // indirect
+	github.com/hashicorp/hcl v1.0.1-vault-5 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/magiconair/properties v1.8.7 // indirect
+	github.com/matttproud/golang_protobuf_extensions v1.0.4 // indirect
+	github.com/mitchellh/mapstructure v1.5.1-0.20220423185008-bf980b35cac4 // indirect
+	github.com/pelletier/go-toml v1.9.5 // indirect
+	github.com/philhofer/fwd v1.1.2 // indirect
+	github.com/pmezard/go-difflib v1.0.1-0.20181226105442-5d4384ee4fb2 // indirect
+	github.com/prometheus/client_golang v1.17.0 // indirect
+	github.com/prometheus/client_model v0.5.0 // indirect
+	github.com/prometheus/common v0.44.0 // indirect
+	github.com/prometheus/procfs v0.12.0 // indirect
+	github.com/spf13/afero v1.9.5 // indirect
+	github.com/spf13/cast v1.5.1 // indirect
+	github.com/spf13/cobra v1.7.0 // indirect
+	github.com/spf13/jwalterweatherman v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+	github.com/stretchr/objx v0.5.1 // indirect
+	github.com/tinylib/msgp v1.1.8 // indirect
+	go.opentelemetry.io/otel v1.20.0 // indirect
+	go.opentelemetry.io/otel/exporters/prometheus v0.42.0 // indirect
+	go.opentelemetry.io/otel/metric v1.20.0 // indirect
+	go.opentelemetry.io/otel/sdk v1.20.0 // indirect
+	go.opentelemetry.io/otel/sdk/metric v1.20.0 // indirect
+	go.opentelemetry.io/otel/trace v1.20.0 // indirect
+	go.uber.org/dig v1.17.0 // indirect
+	go.uber.org/fx v1.18.2 // indirect
+	go.uber.org/multierr v1.11.0 // indirect
+	go.uber.org/zap v1.26.0 // indirect
+	golang.org/x/net v0.18.0 // indirect
+	golang.org/x/sys v0.14.1-0.20231108175955-e4099bfacb8c // indirect
+	golang.org/x/text v0.14.0 // indirect
+	google.golang.org/genproto v0.0.0-20231030173426-d783a09b4405 // indirect
+	google.golang.org/genproto/googleapis/api v0.0.0-20231016165738-49dd2c1f3d0b // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20231106174013-bbf56f31fb17 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
