+++ pkg/remoteconfig/service/service.go
@@ -28,13 +28,13 @@ import (
 	"go.etcd.io/bbolt"
 
 	"github.com/DataDog/datadog-agent/pkg/config"
-	"github.com/DataDog/datadog-agent/pkg/config/remote/api"
-	rdata "github.com/DataDog/datadog-agent/pkg/config/remote/data"
-	"github.com/DataDog/datadog-agent/pkg/config/remote/meta"
-	"github.com/DataDog/datadog-agent/pkg/config/remote/telemetry"
-	"github.com/DataDog/datadog-agent/pkg/config/remote/uptane"
 	configUtils "github.com/DataDog/datadog-agent/pkg/config/utils"
 	pbgo "github.com/DataDog/datadog-agent/pkg/proto/pbgo/core"
+	"github.com/DataDog/datadog-agent/pkg/remoteconfig/api"
+	rdata "github.com/DataDog/datadog-agent/pkg/remoteconfig/data"
+	"github.com/DataDog/datadog-agent/pkg/remoteconfig/meta"
+	"github.com/DataDog/datadog-agent/pkg/remoteconfig/telemetry"
+	"github.com/DataDog/datadog-agent/pkg/remoteconfig/uptane"
 	"github.com/DataDog/datadog-agent/pkg/util/backoff"
 	"github.com/DataDog/datadog-agent/pkg/util/hostname"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
