--- pkg/remoteconfig/state/go.mod
@@ -1,16 +1,0 @@
-module github.com/DataDog/datadog-agent/pkg/remoteconfig/state
-
-go 1.20
-
-require (
-	github.com/DataDog/go-tuf v1.0.2-0.5.2
-	github.com/pkg/errors v0.9.1
-	github.com/secure-systems-lab/go-securesystemslib v0.7.0
-	github.com/stretchr/testify v1.8.4
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
-)
