+++ pkg/remoteconfig/telemetry/README.md
