+++ pkg/remoteconfig/uptane/client_test.go
@@ -19,8 +19,8 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/DataDog/datadog-agent/pkg/config"
-	"github.com/DataDog/datadog-agent/pkg/config/remote/meta"
 	pbgo "github.com/DataDog/datadog-agent/pkg/proto/pbgo/core"
+	"github.com/DataDog/datadog-agent/pkg/remoteconfig/meta"
 )
 
 func getTestOrgUUIDProvider(orgID int) OrgUUIDProvider {
